Remove config
