+++ packages/pixeloven-webpack-dev-server/package.json
@@ -31,7 +31,6 @@
     "express" :"4.16.4",
     "html-entities" :"1.2.1",
     "react-dev-utils" :"6.1.1",
-    "require-from-string" :"2.0.2",
     "webpack": "4.28.2",
     "webpack-dev-middleware": "3.4.0",
     "webpack-hot-middleware": "2.24.3",
@@ -41,7 +40,6 @@
     "@types/express": "4.16.0",
     "@types/html-entities": "1.2.16",
     "@types/node": "10.12.12",
-    "@types/require-from-string" : "1.2.0",
     "@types/webpack": "4.4.22",
     "@types/webpack-dev-middleware": "2.0.2",
     "@types/webpack-hot-middleware": "2.16.4"
