+++ packages/pixeloven-webpack-dev-server/src/lib/Server.ts
@@ -1,17 +1,30 @@
+import { logger } from "@pixeloven/node-logger";
 import express from "express";
+import fs from "fs";
 import path from "path";
-import { ServerConfig } from "./ServerConfig";
+import Compiler from "./Compiler";
+import { Config } from "./config";
+import {
+    createWebpackDevMiddleware,
+    createWebpackHotMiddleware,
+    createWebpackHotServerMiddleware,
+    errorHandler,
+} from "./middleware";
+
 
 type ServerOnComplete = (error?: Error) => void;
 
 class Server {
-    protected config: ServerConfig;
+    protected compiler: Compiler;
+    protected config: Config;
 
     /**
      * Construct server
+     * @param compiler 
      * @param config 
      */
-    constructor(config: ServerConfig) {
+    constructor(compiler: Compiler, config: Config) {
+        this.compiler = compiler;
         this.config = config;
     }
 
@@ -24,25 +37,35 @@ class Server {
         onComplete: ServerOnComplete,
     ) {
         const app = express();
-        if (this.config.before) {
-            this.config.before(app);
+        if (this.compiler.server) {
+            this.compiler.server.hooks.done.tap("PixelOvenWebpackDevServer", () => {
+                logger.info("DONE!!!!!!!!!!!!");
+            });
         }
-        // Register static file serving
+        /**
+         * @todo Support coverage and type docs. Plus storybook docs
+         */
+        const publicPath = path.resolve(process.cwd(), "public");
+        if (fs.existsSync(publicPath)) {
+            logger.info("Static file directory \"public\" found. Automatically setting up static file serving.");
+            app.use(
+                this.config.path,
+                express.static(publicPath),
+            );
+        } // TODO should we log here?
+        app.use(createWebpackDevMiddleware(this.config, this.compiler.combined));
+        if (this.compiler.client) {
+            app.use(createWebpackHotMiddleware(this.config, this.compiler.client));
+        }
         app.use(
-            this.config.server.path,
-            express.static(path.resolve(process.cwd(), "public")), // TODO check it exists first
+            createWebpackHotServerMiddleware(this.config, this.compiler.combined),
         );
-        // TODO we should also host coverage docs and docs if they exist
-        if (this.config.middle) {
-            this.config.middle(app);
-        }
-        if (this.config.after) {
-            this.config.after(app);
-        }
+        app.use(errorHandler);
+
         // Start express server on specific host and port
         app.listen(
-            this.config.server.port,
-            this.config.server.host,
+            this.config.port,
+            this.config.host,
             onComplete,
         );
     }
